requirements.txt
fastapi==0.116.1
uvicorn[standard]
sqlalchemy==2.0.25
pymysql==1.1.2
pydantic==2.11.7
python-multipart==0.0.20
websockets==15.0.1
SpeechRecognition==3.14.3
pydub==0.25.1
faster-whisper==1.2.0
python-docx==1.2.0
reportlab==4.4.3
Pillow==10.2.0
jinja2==3.1.6
aiofiles==24.1.0
email-validator
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
asyncio-mqtt
requests==2.31.0
python-dotenv==1.1.1
fastapi_utils
loguru==0.7.3
pytz==2023.3.post1
pytest==8.4.1
pytest-asyncio==1.1.0
PyAudio==0.2.14
pocketsphinx
asyncmy